Added workaround for #12
Right now, ssh-import-id requires python3-requests that requires a
version of python3-urllib3 below a certain number but there is an
update. Which in turn, becomes a conflict, so a bunch of things get
uninstalled. Blacklisting ssh-import-id (Importing GitHub and Landscape
SSH keys which is useless on this system) is the solution.